--- src/components/DropDown.tsx
@@ -1,14 +1,0 @@
-function DropDown() {
-  return (
-    <>
-      <div>버튼</div>
-      <ul>
-        <li>settings</li>
-        <li>내가 작성한 글</li>
-        <li>로그아웃</li>
-      </ul>
-    </>
-  );
-}
-
-export default DropDown;
