+++ src/components/Header/Index.tsx
@@ -1,14 +1,14 @@
 'use client';
 
+import DropDown from '../DropDown/Index';
 import styles from './Header.module.scss';
-import Hamburger from '../assets/svg/hamburger.svg';
 
 function Header() {
   return (
     <header className={styles.headerContainer}>
       <div className={styles.headerWrapper}>
         <h2 className={styles.title}>TAGLOG</h2>
-        <Hamburger className={styles.DropdownBtn} />
+        <DropDown />
       </div>
     </header>
   );
